% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{build_catalog}
\alias{build_catalog}
\title{Build a catalog object structure}
\usage{
build_catalog(id, title, description)
}
\arguments{
\item{id}{Catalog ID}

\item{title}{Catalog title}

\item{description}{Catalog description}
}
\value{
List with catalog structure (no links)
}
\description{
Build a catalog object structure
}
\keyword{internal}
