% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{build_catalog_links}
\alias{build_catalog_links}
\title{Build links for a root catalog}
\usage{
build_catalog_links(catalog_file, catalog_dir, title)
}
\arguments{
\item{catalog_file}{Path to catalog.json}

\item{catalog_dir}{Path to catalog directory}

\item{title}{Title of this catalog}
}
\value{
List of link objects
}
\description{
Build links for a root catalog
}
\keyword{internal}
