% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{build_collection_links}
\alias{build_collection_links}
\title{Build links for a collection}
\usage{
build_collection_links(collection_dir, parent_path, items_dir)
}
\arguments{
\item{collection_dir}{Path to collection directory}

\item{parent_path}{Path to parent STAC file}

\item{items_dir}{Path to items directory}
}
\value{
List of link objects
}
\description{
Build links for a collection
}
\keyword{internal}
