% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{build_item_links}
\alias{build_item_links}
\title{Build links for a STAC item}
\usage{
build_item_links(item_id, items_dir, collection_dir)
}
\arguments{
\item{item_id}{Item ID}

\item{items_dir}{Path to items directory}

\item{collection_dir}{Path to collection directory}
}
\value{
List of link objects
}
\description{
Build links for a STAC item
}
\keyword{internal}
