% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{Check file names}
\usage{
check_names(
  path,
  prefix = "3dm",
  zone = 32,
  region = "ni",
  date = NULL,
  full.names = FALSE
)
}
\arguments{
\item{path}{A path to a laz file or a directory which contains laz files}

\item{prefix}{3 letter character. Naming prefix (defaults to "3dm")}

\item{zone}{2 digits integer. UTM zone (defaults to 32)}

\item{region}{2 letter character. Region abbreviation (defaults to "ni")}

\item{date}{YYYYMMDD. (optional) acquisition date}

\item{full.names}{Whether to return the full file path or just the file name (default)}
}
\value{
A dataframe with name_is, name_should, correct
}
\description{
Checks the file names according to our own standard.
File names should be in the following schema:
"prefix_UTMzone_minx_miny_tilesize_region_acquisitiondate.laz"
}
\details{
(e.g. "3dm_32_547_5724_1_ni_20240327.laz")
}
\examples{
f <- system.file("extdata", package="managelidar")
check_names(f)
}
