% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{extract_crs}
\alias{extract_crs}
\title{Extract CRS from VPC items}
\usage{
extract_crs(vpc_obj)
}
\arguments{
\item{vpc_obj}{VPC object}
}
\value{
Integer CRS code
}
\description{
Extract CRS from VPC items
}
\keyword{internal}
