% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{extract_spatial_extent}
\alias{extract_spatial_extent}
\title{Extract spatial extent from VPC items}
\usage{
extract_spatial_extent(vpc_obj)
}
\arguments{
\item{vpc_obj}{VPC object}
}
\value{
List with spatial extent structure
}
\description{
Extract spatial extent from VPC items
}
\keyword{internal}
