% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{extract_temporal_extent}
\alias{extract_temporal_extent}
\title{Extract temporal extent from VPC items}
\usage{
extract_temporal_extent(vpc_obj)
}
\arguments{
\item{vpc_obj}{VPC object}
}
\value{
List with temporal extent structure
}
\description{
Extract temporal extent from VPC items
}
\keyword{internal}
