% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{find_link}
\alias{find_link}
\title{Find the first link with a given \code{rel} value}
\usage{
find_link(links, rel)
}
\arguments{
\item{links}{List of link objects (as returned by \code{\link[=read_stac]{read_stac()}})}

\item{rel}{Relation type to search for (e.g. \code{"root"}, \code{"parent"})}
}
\value{
The matching link object, or \code{NULL} if none found
}
\description{
Find the first link with a given \code{rel} value
}
\keyword{internal}
