% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{find_root_catalog}
\alias{find_root_catalog}
\title{Find root catalog by following links}
\usage{
find_root_catalog(stac_path)
}
\arguments{
\item{stac_path}{Path to any STAC file}
}
\value{
Absolute path to root catalog
}
\description{
Find root catalog by following links
}
\keyword{internal}
