% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{get_crs}
\alias{get_crs}
\title{Get the Coordinate Reference system of laz files}
\usage{
get_crs(path, full.names = FALSE)
}
\arguments{
\item{path}{A path to a laz file or a directory which contains laz files}

\item{full.names}{Whether to return the full file path or just the file name (default)}
}
\value{
A dataframe with file and crs (EPSG)
}
\description{
Get the Coordinate Reference system of laz files
}
\examples{
f <- system.file("extdata", package="managelidar")
get_crs(f)
}
