% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datetime.R
\name{get_datetime}
\alias{get_datetime}
\title{Get acquisition datetime from point cloud}
\usage{
get_datetime(
  path,
  full.names = FALSE,
  verbose = FALSE,
  from_csv = NULL,
  return_referenceyear = FALSE
)
}
\arguments{
\item{path}{The path to a file (.las/.laz/.copc), to a directory which contains these files, or to a virtual point cloud (.vpc) referencing these files.}

\item{full.names}{Whether to return the full file paths or just the filenames (default) Whether to return the full file path or just the file name (default)}

\item{from_csv}{When NULL (default) the datetime will be extracted from internal LAS data, if it points to a valid csv file the datetime will be read from this file.}

\item{return_referenceyear}{If TRUE (FALSE is default) and a valid csv file is provided the reference year will be read from file instead of the date (e.g. referenceyear might be 2015 if data was acquired in december 2014).}
}
\value{
A data.frame with attributes \code{filename}, \code{datetime_min} and \code{datetime_max}
}
\description{
\code{get_datetime()} derives the acquisition datetime for LAS files.
}
\details{
This function derives acquistion dates for LAS files either from inheritted data or from an external csv file. Deriving dates from internal data requires the entire point cloud to be read! Thus it takes potentially a long processing time, especially if applied on entire folders or large files.
Unfortunately it is only possible to derive exact acquisition date from newer (>= LAS 1.3) point cloud data, where acquisition date is encoded as adjusted standard GPS time. Otherwise, additional information on GPSweek is necessary. If \code{from_csv} points to a csv file with a date column and minx/miny in the same CRS than the LAS data, the date is extracted here. inst/extdata/acquisition_dates_lgln.csv contains dates for federal data in Lower Saxony.
}
\examples{
f <- system.file("extdata", package = "managelidar")
get_datetime(f)
}
