% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent.R
\name{get_extent}
\alias{get_extent}
\title{Get the spatial extent of LAS files}
\usage{
get_extent(path, as_sf = FALSE, full.names = FALSE)
}
\arguments{
\item{path}{Character. Path to a LAS/LAZ/COPC file, a directory, or a Virtual Point Cloud (.vpc) referencing these files.}

\item{as_sf}{Logical. If \code{TRUE}, returns an \code{sf} object with geometry.}

\item{full.names}{Logical. If \code{TRUE}, filenames in the output are full paths; otherwise base filenames (default).}
}
\value{
A \code{data.frame} or \code{sf} object with columns:
\describe{
\item{filename}{Filename of the LAS file.}
\item{xmin}{Minimum X coordinate.}
\item{xmax}{Maximum X coordinate.}
\item{ymin}{Minimum Y coordinate.}
\item{ymax}{Maximum Y coordinate.}
\item{geometry}{(optional) Polygon geometry if \code{as_sf = TRUE}.}
}
}
\description{
\code{get_extent()} extracts the spatial extent (xmin, xmax, ymin, ymax) from LASfiles.
}
\examples{
folder <- system.file("extdata", package = "managelidar")
las_files <- list.files(folder, full.names = T, pattern = "*20240327.laz")
las_files |> get_extent()

}
