% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extrabytes_attributes.R
\name{get_extrabytes_attributes}
\alias{get_extrabytes_attributes}
\title{Get the extrabytes attributes stored in LASfiles}
\usage{
get_extrabytes_attributes(path, full.names = FALSE)
}
\arguments{
\item{path}{Character. Path(s) to LAS/LAZ/COPC files, a directory containing
such files, or a Virtual Point Cloud (.vpc) referencing these files.}

\item{full.names}{Logical. If \code{TRUE}, filenames in the output are full paths;
otherwise base filenames (default).}
}
\value{
A \code{data.frame} with columns:
\describe{
\item{filename}{Filename of the LASfile.}
\item{extrabytes_attributes}{list of attribute names}
}
}
\description{
\code{get_extrabytes_attributes()} extracts the names of extrabytes attributes
stored in LASfiles.
}
\examples{
folder <- system.file("extdata", package = "managelidar")
las_files <- list.files(folder, full.names = T, pattern = "*20240327.laz")

folder |> get_extrabytes_attributes()

}
