% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gpkg.R
\name{get_gpkg}
\alias{get_gpkg}
\title{Create a Geopackage containing metadata of LAS files}
\usage{
get_gpkg(
  path,
  out_file = tempfile(fileext = ".gpkg"),
  overwrite = FALSE,
  crs = 25832,
  metrics = NULL
)
}
\arguments{
\item{path}{Character. Path to a LAS/LAZ/COPC file, a directory containing LAS files,
or a Virtual Point Cloud (.vpc) file.}

\item{out_file}{Path to the output Geopackage (.gpkg) file (default: tempfile).}

\item{overwrite}{Logical. If TRUE, overwrite the output file if it exists (default: FALSE).}

\item{crs}{Integer. Optional EPSG code to reproject the VPC (default: 25832).}
}
\value{
Invisibly returns an \code{sf} object representing the tiles written to the Geopackage.
}
\description{
\code{get_gpkg()} converts the metadata of a Virtual Point Cloud (.vpc) or a collection of LAS/LAZ/COPC
files into Geopackage. VPCs can be read and vizualized by QGIS, however individual tiles (features) can not be queried as is.
To do this we convert it to a Geopackage, which can easily be explored in any GIS.
Each LAS tile becomes a feature with its spatial extent and some metadata.
}
\details{
Summary metrics (optional) can be included by setting \code{metrics = TRUE} for default metrics
or providing a character vector of custom metrics. Computing metrics requires reading the actual point data,
und thus can be much slower. See \code{get_summary()} for details.
}
\examples{
f <- system.file("extdata", package = "managelidar")
get_gpkg(f, tempfile(fileext = ".gpkg"))

}
