% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersection.R
\name{get_intersection}
\alias{get_intersection}
\title{Get intersecting LAS files}
\usage{
get_intersection(
  path1,
  path2,
  mode = "intersects",
  as_sf = FALSE,
  full.names = FALSE
)
}
\arguments{
\item{path1}{The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.}

\item{path2}{The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.}

\item{mode}{Either 'intersects' (default) or 'equals'.}

\item{as_sf}{(boolean) Whether to return the dataframe as spatials features data.frame or not (default).}

\item{full.names}{Whether to return the full file paths or just the filenames (default)}
}
\value{
A list of 2 data.frames with attribute \code{filename} of all intersecting or equal file extents between \code{path1} and \code{path2}, optionally as spatial features data.frame.
}
\description{
Get intersecting LAS files
}
\examples{
folder <- system.file("extdata", package = "managelidar")
file <- list.files(folder, full.names = T)[1]
get_intersection(folder, file)
}
