% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{get_items_dir}
\alias{get_items_dir}
\title{Get items directory path}
\usage{
get_items_dir(collection_dir)
}
\arguments{
\item{collection_dir}{Path to collection directory}
}
\value{
Path to items directory
}
\description{
Get items directory path
}
\keyword{internal}
