% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lasversion.R
\name{get_lasversion}
\alias{get_lasversion}
\title{Get the Version of LAS files}
\usage{
get_lasversion(path, full.names = FALSE)
}
\arguments{
\item{path}{A path to a LAS file or a directory which contains LAS files}

\item{full.names}{Whether to return the full file paths or just the filenames (default) Whether to return the full file path or just the file name (default)}
}
\value{
A data.frame with attributes \code{filename} and \code{lasversion} (Major.Minor)
}
\description{
Get the Version of LAS files
}
\examples{
f <- system.file("extdata", package="managelidar")
get_lasversion(f)
}
