% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get file names}
\usage{
get_names(path, full.names = FALSE)
}
\arguments{
\item{path}{A path to a LAS file, VPC file, or a directory which contains LAS files}

\item{full.names}{Whether to return the full file paths or just the filenames (default) Whether to return the full file path or just the file name (default)}
}
\value{
A vector of filenames
}
\description{
Simply get a vector of names of all LAS files (*.las, *.laz, *.laz.copc) in a folder or VPC.
}
\examples{
f <- system.file("extdata", package = "managelidar")
get_names(f)
}
