% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{get_stac_type}
\alias{get_stac_type}
\title{Get STAC object type}
\usage{
get_stac_type(stac_obj)
}
\arguments{
\item{stac_obj}{STAC object (list)}
}
\value{
Character: "Catalog" or "Collection"
}
\description{
Get STAC object type
}
\keyword{internal}
