% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{Any R objects.}
}
\value{
\code{x} if non-\code{NULL}, otherwise \code{y}.
}
\description{
Returns \code{x} if not \code{NULL}, otherwise \code{y}.
}
\keyword{internal}
