% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_classified.R
\name{is_classified}
\alias{is_classified}
\title{Check if LAS files are classified}
\usage{
is_classified(path, full.names = FALSE, add_classes = FALSE)
}
\arguments{
\item{path}{The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.}

\item{full.names}{Whether to return the full file paths or just the filenames (default) Whether to return the full file paths or just the filenames (default)}

\item{add_classes}{Whether to add a list of present classes or not (default)}
}
\value{
A data.frame returning \code{filename}, \code{classified} and optionally \code{classes}
}
\description{
\code{is_classified} derives information about point classification
}
\details{
The function needs to read the actual point cloud data!
To speed up the processing the function reads just a sample of points, which is slower than just reading the header information but much faster than reading the entire file.
The results are thus only valid for the subsample of points and do not necessarily reflect the entire file.
}
\examples{
f <- system.file("extdata", package = "managelidar")
is_classified(f)

}
