% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_indexed.R
\name{is_indexed}
\alias{is_indexed}
\title{Check whether LASfiles are spatially indexed}
\usage{
is_indexed(path, full.names = FALSE)
}
\arguments{
\item{path}{Character. Path(s) to a LAS/LAZ/COPC file, a directory containing
such files, or a Virtual Point Cloud (\code{.vpc}).}

\item{full.names}{Logical. If \code{TRUE}, return full file paths; otherwise
return base filenames only (default).}
}
\value{
A \code{data.frame} with columns:
\describe{
\item{file}{Filename of the LASfile.}
\item{is_indexed}{Logical indicating whether point cloud is spatially indexed}
}
}
\description{
\code{is_indexed()} whether LASfiles are spatially indexed (either via external \code{.lax} file or internally)
}
\details{
The input may be a single file, a directory containing LASfiles, or a
Virtual Point Cloud (\code{.vpc}) referencing LAS/LAZ/COPC files. Internally,
file paths are resolved using \code{resolve_las_paths()}.
}
\examples{
folder <- system.file("extdata", package = "managelidar")
las_files <- list.files(folder, full.names = T, pattern = "*20240327.laz")

las_files |> is_indexed()

}
