% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laz_files.R
\docType{data}
\name{laz_files}
\alias{laz_files}
\title{Laz files}
\format{
LAZ files
}
\source{
flight campaign "Solling 2024" (local data)
}
\usage{
laz_files
}
\description{
4 neighbouring laz tiles of an ALS campaign, thinned to 2m grid
}
\keyword{datasets}
