% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managelidar-package.R
\docType{package}
\name{managelidar-package}
\alias{managelidar}
\alias{managelidar-package}
\title{managelidar: Manage Lidar Data Files}
\description{
The goal of managelidar is to facilitate the handling of lidar data files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wiesehahn/managelidar}
  \item \url{https://wiesehahn.github.io/managelidar/}
  \item Report bugs at \url{https://github.com/wiesehahn/managelidar/issues}
}

}
\author{
\strong{Maintainer}: Jens Wiesehahn \email{wiesehahn.jens@gmail.com}

}
\keyword{internal}
