% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{merge_spatial_extents}
\alias{merge_spatial_extents}
\title{Merge two spatial extents}
\usage{
merge_spatial_extents(extent1, extent2)
}
\arguments{
\item{extent1}{Spatial extent list (existing, bbox as matrix from yyjsonr)}

\item{extent2}{Spatial extent list (new, bbox as list)}
}
\value{
Merged spatial extent list
}
\description{
Merge two spatial extents
}
\keyword{internal}
