% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{merge_temporal_extents}
\alias{merge_temporal_extents}
\title{Merge two temporal extents}
\usage{
merge_temporal_extents(extent1, extent2)
}
\arguments{
\item{extent1}{Temporal extent list (existing, interval as matrix from yyjsonr)}

\item{extent2}{Temporal extent list (new, interval as list)}
}
\value{
Merged temporal extent list
}
\description{
Merge two temporal extents
}
\keyword{internal}
