% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_temporal.R
\name{parse_datetime_input}
\alias{parse_datetime_input}
\title{Internal helper to parse datetime input}
\usage{
parse_datetime_input(dt, position = "start")
}
\arguments{
\item{dt}{POSIXct, Date, or character datetime}

\item{position}{Either "start" or "end" - determines whether to use beginning or end of period}
}
\value{
POSIXct object in UTC
}
\description{
Internal helper to parse datetime input
}
\keyword{internal}
