% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extent.R
\name{plot_extent}
\alias{plot_extent}
\title{Plot the spatial extent of LAS files}
\usage{
plot_extent(path, full.names = FALSE)
}
\arguments{
\item{path}{Character. Path(s) to LAS/LAZ/COPC files, a directory containing
such files, or a Virtual Point Cloud (.vpc).}
}
\value{
An interactive \code{mapview} map displayed in the viewer.
}
\description{
\code{plot_extent()} visualizes the spatial extent of LAS/LAZ/COPC files on an
interactive map using bounding boxes derived from file headers or an
existing Virtual Point Cloud (VPC).
}
\examples{
f <- system.file("extdata", package = "managelidar")
plot_extent(f)
}
