% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{read_stac}
\alias{read_stac}
\title{Read a STAC JSON file}
\usage{
read_stac(path)
}
\arguments{
\item{path}{Path to STAC JSON file}
}
\value{
List representing STAC object
}
\description{
Read a STAC JSON file
}
\keyword{internal}
