% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{required_lidar_stac_extensions}
\alias{required_lidar_stac_extensions}
\title{STAC extensions required whenever pointcloud/projection fields
(\code{proj:epsg}, \code{pc:type}) are present on a collection's summaries.}
\usage{
required_lidar_stac_extensions()
}
\value{
Character vector of extension schema URLs
}
\description{
STAC extensions required whenever pointcloud/projection fields
(\code{proj:epsg}, \code{pc:type}) are present on a collection's summaries.
}
\keyword{internal}
