% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{resolve_collection_dir}
\alias{resolve_collection_dir}
\title{Resolve collection directory based on parent type}
\usage{
resolve_collection_dir(parent_path, collection_id)
}
\arguments{
\item{parent_path}{Path to parent STAC file}

\item{collection_id}{Collection ID}
}
\value{
Path to collection directory (created if needed)
}
\description{
Resolve collection directory based on parent type
}
\keyword{internal}
