% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{set_names}
\alias{set_names}
\title{Set file names}
\usage{
set_names(path, prefix = "3dm", zone = 32, region = "ni", date = NULL)
}
\arguments{
\item{path}{Path to folder containing laz files}
}
\value{
Renamed files
}
\description{
Renames files according to schema validated by \code{check_files()}.
}
\examples{
f <- system.file("extdata/3dm_32_547_5724_1_ni_20240327.laz", package="managelidar")
tmpname <- file.path(dirname(f), "wrongname.laz")
file.copy(f, tmpname)
set_names(tmpname)
file.remove(tmpname)
}
