% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_browse.R
\name{stac_browse_handle_request}
\alias{stac_browse_handle_request}
\title{Handle a single HTTP request for stac_browse()}
\usage{
stac_browse_handle_request(req, root_dir)
}
\arguments{
\item{req}{httpuv request object}

\item{root_dir}{Root directory to serve files from}
}
\value{
httpuv response list
}
\description{
Handle a single HTTP request for stac_browse()
}
\keyword{internal}
