% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{stac_timestamp}
\alias{stac_timestamp}
\title{Current time as an RFC 3339 / STAC-compliant UTC timestamp}
\usage{
stac_timestamp()
}
\value{
Character, e.g. \code{"2026-07-15T14:32:07Z"}
}
\description{
Current time as an RFC 3339 / STAC-compliant UTC timestamp
}
\keyword{internal}
