% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{write_items}
\alias{write_items}
\title{Write STAC items to files}
\usage{
write_items(items, items_dir, overwrite = FALSE)
}
\arguments{
\item{items}{List of STAC item objects}

\item{items_dir}{Path to items directory}

\item{overwrite}{Logical. If FALSE (default), skip existing items}
}
\value{
Character vector of written item IDs
}
\description{
Write STAC items to files
}
\keyword{internal}
