% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{write_stac}
\alias{write_stac}
\title{Write a STAC object to JSON file}
\usage{
write_stac(obj, path)
}
\arguments{
\item{obj}{STAC object (list)}

\item{path}{Output path}
}
\value{
Invisible NULL
}
\description{
Write a STAC object to JSON file
}
\keyword{internal}
