% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{add_child_link}
\alias{add_child_link}
\title{Add a child link to parent object}
\usage{
add_child_link(parent_obj, child_rel_path, child_title = NULL)
}
\arguments{
\item{parent_obj}{Parent STAC object (list)}

\item{child_rel_path}{Relative path to child from parent directory}

\item{child_title}{Optional title for the link}
}
\value{
Updated parent object
}
\description{
Add a child link to parent object
}
\keyword{internal}
