% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{Check file names}
\usage{
check_names(
  path,
  prefix = "3dm",
  zone = 32,
  region = NULL,
  year = NULL,
  copc = FALSE,
  full.names = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{path}{A path to a directory which contains las/laz files}

\item{prefix}{3 letter character. Naming prefix (defaults to "3dm")}

\item{zone}{2 digits integer. UTM zone (defaults to 32)}

\item{region}{2 letter character. (optional) federal state abbreviation. It will be fetched automatically if not defined (default).}

\item{year}{YYYY. (optional) acquisition year to append to filename.
If not provided (default) the year will be extracted from the files. It will be the acquisition date if points contain datetime in GPStime format, otherwise it will get the year from the file header, which is the processing date by definition.}

\item{copc}{Whether the file is expected to be a Cloud Optimized Point Cloud (.copc.laz)}

\item{full.names}{Whether to return the full file path or just the file names (default)}
}
\value{
A dataframe with name_is, name_should, correct
}
\description{
Checks the file names according to \href{https://www.adv-online.de/AdV-Produkte/Standards-und-Produktblaetter/Standards-der-Geotopographie/binarywriterservlet?imgUid=6b510f6e-a708-d081-505a-20954cd298e1&uBasVariant=11111111-1111-1111-1111-111111111111}{ADV standard}.
File names should be in the following schema:
\code{prefix}\emph{\code{utmzone}}\code{minx}\emph{\code{miny}}\code{tilesize}\emph{\code{region}}\verb{year``.laz}
}
\details{
(e.g. \verb{3dm_32_547_5724_1_ni_2024.laz})
}
\examples{
f <- system.file("extdata", package = "managelidar")
check_names(f)
}
