% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{get_crs}
\alias{get_crs}
\title{Get the Coordinate Reference System (CRS) of LAS files}
\usage{
get_crs(path, full.names = FALSE)
}
\arguments{
\item{path}{Character. Path to a LAS/LAZ/COPC file, a directory containing LAS files,
or a Virtual Point Cloud (.vpc) referencing LAS files.}

\item{full.names}{Logical. If \code{TRUE}, the returned filenames will be full paths;
if \code{FALSE} (default), only base filenames are used.}
}
\value{
A \code{data.frame} with two columns:
\describe{
\item{filename}{The filename or full path of each LAS file.}
\item{crs}{The EPSG code of the file's coordinate reference system.}
}
}
\description{
\code{get_crs()} efficiently extracts and returns the coordinate reference system (EPSG code) of LAS files.
}
\details{
This function efficiently reads the Coordinate Reference System of LAS files from VPC. It is suitable
for quickly inspecting the CRS of multiple LAS/LAZ/COPC files.
}
\examples{
f <- system.file("extdata", package = "managelidar")
get_crs(f)

}
