% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{get_crs}
\alias{get_crs}
\title{Get the Coordinate Reference System of LAS files}
\usage{
get_crs(path, full.names = FALSE)
}
\arguments{
\item{path}{The path to a file (.las/.laz/.copc), to a directory which contains these files, or to a virtual point cloud (.vpc) referencing these files.}

\item{full.names}{Whether to return the full file path or just the file name (default)}
}
\value{
A dataframe returning \code{filename} and \code{crs} (EPSG)
}
\description{
Get the Coordinate Reference System of LAS files
}
\examples{
f <- system.file("extdata", package="managelidar")
get_crs(f)

}
