% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datetime.R
\name{get_datetime}
\alias{get_datetime}
\title{Get acquisition datetime from point cloud}
\usage{
get_datetime(path, full.names = FALSE, verbose = FALSE)
}
\arguments{
\item{path}{A path to a laz file or a directory which contains laz files}

\item{full.names}{Whether to return the full file path or just the file name (default)}
}
\value{
A dataframe with file, min datetime, max datetime
}
\description{
\code{get_datetime()} calculates the earliest and latest acquisition datetime from point cloud files.
}
\details{
Unfortunately this is only possible for newer (>= LAS 1.3) point cloud data, where acquisition date is encoded as adjusted standard GPS time. This can be converted to datetime. This function requires the entire point cloud to be read and thus takes long processing time, especially if applied on entire folders or large files.
}
\examples{
f <- system.file("extdata", package = "managelidar")
get_datetime(f)
}
