% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datetime.R
\name{get_datetime}
\alias{get_datetime}
\title{Get acquisition datetime from point cloud}
\usage{
get_datetime(path, full.names = FALSE, verbose = FALSE)
}
\arguments{
\item{path}{The path to a file (.las/.laz/.copc), to a directory which contains these files, or to a virtual point cloud (.vpc) referencing these files.}

\item{full.names}{Whether to return the full file path or just the file name (default)}
}
\value{
A dataframe returning \code{filename}, \code{datetime_min} \code{datetime_max}
}
\description{
\code{get_datetime()} calculates the earliest and latest acquisition datetime from point cloud files.
}
\details{
This function requires the entire point cloud to be read! Thus it takes potentially a long processing time, especially if applied on entire folders or large files.
Unfortunately it is only possible to derive exact acquisition date from newer (>= LAS 1.3) point cloud data, where acquisition date is encoded as adjusted standard GPS time. Otherwise, additional information on GPSweek is necessary.
}
\examples{
f <- system.file("extdata", package = "managelidar")
get_datetime(f)
}
