% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.R
\name{get_density}
\alias{get_density}
\title{Get approximate point and pulse density of LAS files}
\usage{
get_density(path, full.names = FALSE)
}
\arguments{
\item{path}{Character. Path to a LAS/LAZ/COPC file, a directory containing LAS files,
or a Virtual Point Cloud (.vpc) referencing these files.}

\item{full.names}{Logical. If \code{TRUE}, filenames in the output are full paths; otherwise base filenames (default).}
}
\value{
A \code{data.frame} with columns:
\describe{
\item{filename}{File name or path.}
\item{npoints}{Total number of points in the file.}
\item{npulses}{Number of first-return pulses.}
\item{area}{Area of bounding box (units of CRS^2).}
\item{pointdensity}{Approximate points per unit area.}
\item{pulsedensity}{Approximate first-return pulses per unit area.}
}
}
\description{
\code{get_density()} calculates the approximate average point and pulse density of LAS files.
}
\details{
Only the LAS file headers are read. Densities are calculated based on the bounding box
and number of points / first-return pulses. This does not account for missing data
within the bounding box, so the density is approximate and faster to compute than reading
the full point cloud.
}
\examples{
folder <- system.file("extdata", package = "managelidar")
las_files <- list.files(folder, full.names = T, pattern = "*20240327.laz")
las_files |> get_density()

}
