% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent.R
\name{get_extent}
\alias{get_extent}
\title{Get the spatial extent of laz files}
\usage{
get_extent(path)
}
\arguments{
\item{path}{Either a path to a directory which contains laz files or
the path to a Virtual Point Cloud (.vpc) created with lasR package.}
}
\value{
A dataframe with file, minx, miny, maxx, maxy
}
\description{
\code{get_extent} uses min and max values of spatial extent defined in the header of lasfiles.
}
\examples{
f <- system.file("extdata", package="managelidar")
get_extent(f)
}
