% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_header.R
\name{get_header}
\alias{get_header}
\title{Get Fileheader from las file}
\usage{
get_header(file)
}
\arguments{
\item{file}{A path to a laz file}
}
\value{
LASheader
}
\description{
Provides a simple wrapper to read the fileheader from LAS/LAZ files via lidR::readLASheader.
}
\examples{
f <- system.file("extdata/3dm_32_547_5724_1_ni_20240327.laz", package="managelidar")
get_header(f)
}
