% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersection.R
\name{get_intersection}
\alias{get_intersection}
\title{Get intersecting LAS files}
\usage{
get_intersection(path1, path2, mode = "intersects", full.names = FALSE)
}
\arguments{
\item{path1}{path The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.}

\item{path2}{The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.}

\item{mode}{Either 'intersects' (default) or 'equals'.}

\item{full.names}{Whether to return the full file paths or just the filenames (default)}
}
\value{
A spatial features data.frame with attributes \code{filename}, \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax} of all files in \code{path1} which intersect files in \code{path2} ('intersects') or which have the same extent as files in \code{path2} ('equals').
}
\description{
Get intersecting LAS files
}
\examples{
folder <- system.file("extdata", package = "managelidar")
file <- list.files(folder, full.names = T)[1]
get_intersection(folder, file)
}
