% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersection.R
\name{get_intersection}
\alias{get_intersection}
\title{Get intersecting LASfiles}
\usage{
get_intersection(
  path1,
  path2,
  mode = "intersects",
  as_sf = FALSE,
  full.names = FALSE
)
}
\arguments{
\item{path1}{Character. Path(s) to LAS/LAZ/COPC files, a directory, or a
Virtual Point Cloud (.vpc).}

\item{path2}{Character. Path(s) to LAS/LAZ/COPC files, a directory, or a
Virtual Point Cloud (.vpc).}

\item{mode}{Character. Spatial predicate to use: \code{"intersects"} (default)
or \code{"equals"}.}

\item{as_sf}{Logical. If \code{TRUE}, return results as \code{sf} objects;
otherwise drop geometries (default).}

\item{full.names}{Logical. If \code{TRUE}, filenames are returned as full paths;
otherwise base filenames (default).}
}
\value{
A named list with two elements (\code{path1}, \code{path2}), each containing
a \code{data.frame} or \code{sf} object with column \code{filename} for intersecting
or equal file extents.
}
\description{
\code{get_intersection()} identifies LAS/LAZ/COPC files whose spatial extents
intersect or are spatially equal between two inputs.
}
\details{
This function simply checks for intersection between two inputs. \code{\link{is_multitemporal}}
in contrast is a newer addition and works with a single input (which can be a vector of multiple files/folders),
in most cases \code{\link{filter_first}} / \code{\link{filter_latest}} might be the best choice.
}
\examples{
folder <- system.file("extdata", package = "managelidar")
las_files <- list.files(folder, full.names = T, pattern = "*20240327.laz")
las_file <- list.files(folder, full.names = T, pattern = "*20230904.laz")
get_intersection(las_files, las_file)
}
