% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lasversion.R
\name{get_lasversion}
\alias{get_lasversion}
\title{Get the Version of laz files}
\usage{
get_lasversion(path, full.names = FALSE)
}
\arguments{
\item{path}{A path to a laz file or a directory which contains laz files}

\item{full.names}{Whether to return the full file path or just the file name (default)}
}
\value{
A dataframe with file and lasversion (Major.Minor)
}
\description{
Get the Version of laz files
}
\examples{
f <- system.file("extdata", package="managelidar")
get_lasversion(f)
}
