% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get LASfile names}
\usage{
get_names(path, full.names = FALSE)
}
\arguments{
\item{path}{Character. Path(s) to a LAS/LAZ/COPC file, a directory containing
such files, or a Virtual Point Cloud (\code{.vpc}).}

\item{full.names}{Logical. If \code{TRUE}, return full file paths; otherwise
return base filenames only (default).}
}
\value{
A character vector of filenames or file paths.
}
\description{
\code{get_names()} returns the filenames of all LAS-related point cloud files
(\code{.las}, \code{.laz}, \code{.copc}) found in a given input.
}
\details{
The input may be a single file, a directory containing LASfiles, or a
Virtual Point Cloud (\code{.vpc}) referencing LAS/LAZ/COPC files. Internally,
file paths are resolved using \code{resolve_las_paths()}.
}
\examples{
folder <- system.file("extdata", package = "managelidar")
las_files <- list.files(folder, full.names = T, pattern = "*20240327.laz")

las_files |> get_names()

}
