% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get file names}
\usage{
get_names(path)
}
\arguments{
\item{path}{A path to a directory which contains laz files}
}
\value{
A vector of filenames
}
\description{
Simply list the names of all lasfiles (*.laz) in the folder.
}
\examples{
f <- system.file("extdata", package="managelidar")
get_names(f)
}
