% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Get the point cloud summary of LAS files}
\usage{
get_summary(path, full.names = FALSE)
}
\arguments{
\item{path}{The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.}

\item{full.names}{Whether to return the full file paths or just the filenames (default) Whether to return the full file path or just the file name (default).}
}
\value{
A named list of summary information (\code{npoints}, \code{nsingle}, \code{nwithheld}, \code{nsynthetic}, \code{npoints_per_return}, \code{npoints_per_class}, \code{z_histogram}, \code{i_histogram}, \code{crs}, \code{epsg})
}
\description{
\code{get_summary} derives information for LAS files, such as number of points per class and histogram distribution of z or intensity values.
}
\details{
The function needs to read the actual point cloud data!
To speed up the processing the function reads just a sample of points, which is slower than just reading the header information but much faster than reading the entire file.
But the results are thus only valid for the subsample of points and do not necessarily reflect the entire file.
}
\examples{
f <- system.file("extdata", package = "managelidar")
get_summary(f)

}
