% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extent.R
\name{plot_extent}
\alias{plot_extent}
\title{Plot the extent of LAS files}
\usage{
plot_extent(path)
}
\arguments{
\item{path}{The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.}
}
\value{
An interactive map in the viewer
}
\description{
\code{plot_extent} plots the spatial extent of LAS files on an interactive map.
}
\examples{
f <- system.file("extdata", package = "managelidar")
plot_extent(f)
}
