% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{set_names}
\alias{set_names}
\title{Set file names}
\usage{
set_names(path, prefix = "3dm", zone = 32, region = "ni", year = NULL)
}
\arguments{
\item{path}{A path to a directory which contains las/laz files}

\item{prefix}{3 letter character. Naming prefix (defaults to "3dm")}

\item{zone}{2 digits integer. UTM zone (defaults to 32)}

\item{region}{2 letter character. Region abbreviation (defaults to "ni")}

\item{year}{YYYY. (optional) acquisition year to append to filename.
If not provided the year will be extracted from the files. It will be the acquisition date if points contain datetime in GPStime format, otherwise it will get the year from the file header, which is the processing date by definition.}
}
\value{
Renamed files
}
\description{
Renames files according to schema validated by \code{check_files()}. This is according to \href{https://www.adv-online.de/AdV-Produkte/Standards-und-Produktblaetter/Standards-der-Geotopographie/binarywriterservlet?imgUid=6b510f6e-a708-d081-505a-20954cd298e1&uBasVariant=11111111-1111-1111-1111-111111111111}{ADV standard}.
}
\examples{
f <- system.file("extdata/3dm_32_547_5724_1_ni_20240327.laz", package = "managelidar")
copy <- tempfile(fileext = ".laz")
file.copy(f, copy)
set_names(copy)
}
