% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{set_names}
\alias{set_names}
\title{Set LASfile names according to ADV standard}
\usage{
set_names(
  path,
  prefix = "3dm",
  zone = 32,
  region = NULL,
  year = NULL,
  copc = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{path}{Character. Path to a LAS/LAZ/COPC file or a directory containing LASfiles.}

\item{prefix}{Character. Naming prefix (default \code{"3dm"}).}

\item{zone}{Integer. UTM zone (default \code{32}).}

\item{region}{Character. Federal state abbreviation (optional). Automatically determined if \code{NULL}.}

\item{year}{Integer or character. Acquisition year to append to filenames (optional). If \code{NULL}, the year is derived from the file.}

\item{copc}{Logical. Whether files are COPC (\code{.copc.laz}, default \code{FALSE}).}

\item{dry_run}{Logical. If \code{TRUE}, only preview renaming without modifying files (default \code{FALSE}).}

\item{verbose}{Logical. Print messages and a preview of renaming (default \code{FALSE}).}
}
\value{
Invisibly returns a \code{data.frame} with columns \code{from} and \code{to} showing original and new filenames.
}
\description{
\code{set_names()} renames LAS/LAZ/COPC files to match the \href{https://www.adv-online.de/AdV-Produkte/Standards-und-Produktblaetter/Standards-der-Geotopographie/binarywriterservlet?imgUid=6b510f6e-a708-d081-505a-20954cd298e1&uBasVariant=11111111-1111-1111-1111-111111111111}{ADV naming standard}.
}
\details{
Files are expected to follow the schema:
\code{prefix_utmzone_minx_miny_tilesize_region_year.laz}
}
\examples{
f <- system.file("extdata", package = "managelidar")
set_names(copy, verbose = TRUE, dry_run = TRUE)

}
