% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_helpers.R
\name{vpc_to_stac_items}
\alias{vpc_to_stac_items}
\title{Convert VPC features to STAC items}
\usage{
vpc_to_stac_items(vpc_obj, collection_dir, items_dir)
}
\arguments{
\item{vpc_obj}{VPC object (list with $features)}

\item{collection_dir}{Path to collection directory}

\item{items_dir}{Path to items directory}
}
\value{
List of STAC item objects
}
\description{
Convert VPC features to STAC items
}
\keyword{internal}
